requirements.txt
absl-py==0.12.0
albumentations @ git+https://github.com/albu/albumentations@5a3b2301f6ebfbece7a2e8c1e134fb08cfdfac86
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==21.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.4.1
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.3
chex==0.0.8
click==8.0.4
cmake==3.22.3
colorama==0.4.4
cycler==0.10.0
Cython==0.29.28
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
dlib==19.23.1
dm-haiku==0.0.4
dm-tree==0.1.6
docker-pycreds==0.4.0
editdistance==0.6.0
entrypoints==0.3
ffmpeg-bitrate-stats==0.2.2
filelock==3.4.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.32.0
h5py==2.10.0
hydra-core==1.0.7
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.6.1
importlib-resources==5.2.0
imutils==0.5.4
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jax==0.2.17
jaxlib==0.1.68+cuda101
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonschema==3.2.0
kaleido==0.2.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==11.1.0
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
mediapipe==0.8.3
mistune==0.8.4
moviepy==1.0.3
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.5
numba==0.53.1
numpy==1.19.3
oauthlib==3.1.1
omegaconf==2.0.6
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
optax==0.0.9
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkg_resources==0.0.0
pooch==1.5.2
portalocker==2.4.0
proglog==0.1.9
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.25.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
sacrebleu==2.0.0
scikit-image==0.17.2
scikit-learn==0.24.2
scikit-video==1.1.11
scipy==1.4.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.9.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1
six==1.15.0
sklearn==0.0
smmap==5.0.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
tabulate==0.8.9
tb-nightly==2.6.0a20210806
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-datasets==4.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-io==0.16.0
tensorflow-metadata==1.1.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tf-estimator-nightly==2.7.0.dev2021080901
threadpoolctl==2.2.0
tifffile==2020.9.3
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.5.0